Some fixes for AV-GS-PP link

+++ SW4STM32/BellaLui/Application/HostBoard/Inc/misc/datastructs.h
@@ -41,11 +41,11 @@ typedef struct
 {
   uint16_t pressure1;
   uint16_t pressure2;
-  uint16_t temperature1;
-  uint16_t temperature2;
-  uint16_t temperature3;
+  int16_t temperature1;
+  int16_t temperature2;
+  int16_t temperature3;
   uint16_t status;
-  uint16_t motor_position;
+  int32_t motor_position;
 } PropulsionData;
 
 typedef struct
